Move filter classes to `filter` package
