+++ sonar-javascript-plugin/src/main/java/org/sonar/plugins/javascript/TypeScriptLanguage.java
@@ -26,8 +26,8 @@ public class TypeScriptLanguage extends AbstractLanguage {
 
   public static final String KEY = "ts";
 
-  static final String FILE_SUFFIXES_KEY = "sonar.typescript.file.suffixes";
-  static final String FILE_SUFFIXES_DEFVALUE = ".ts,.tsx,.cts,.mts";
+  public static final String FILE_SUFFIXES_KEY = "sonar.typescript.file.suffixes";
+  public static final String FILE_SUFFIXES_DEFVALUE = ".ts,.tsx,.cts,.mts";
 
   private Configuration configuration;
