+++ sonar-javascript-plugin/src/main/java/org/sonar/plugins/javascript/filter/MinificationAssessor.java
@@ -17,7 +17,7 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonar.plugins.javascript.minify;
+package org.sonar.plugins.javascript.filter;
 
 import org.sonar.api.batch.fs.InputFile;
 import org.sonar.api.utils.log.Logger;
