+++ sonar-javascript-plugin/src/main/java/org/sonar/plugins/javascript/filter/package-info.java
@@ -17,4 +17,4 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonar.plugins.javascript.minify;
+package org.sonar.plugins.javascript.filter;
