+++ sonar-javascript-plugin/src/test/java/org/sonar/plugins/javascript/filter/MinificationAssessorTest.java
@@ -17,7 +17,7 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonar.plugins.javascript.minify;
+package org.sonar.plugins.javascript.filter;
 
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Paths;
